modify : refactor directory structure
